fix(web): choose environment chat channel
