+++ apps/web/app/design/components-content.tsx
@@ -868,11 +868,18 @@ export function ComponentsContent() {
               will build without showing empty scores or missing facts.
             </p>
             <EnvironmentEmptyState
-              contactAction={{
-                href: "sms:+15555550100?body=I%20want%20to%20update%20what%20you%20know%20about%20my%20home%20environment.",
-                kind: "text",
-                label: "Text Murph",
-              }}
+              contactOptions={[
+                {
+                  href: "sms:+15555550100?body=I%20want%20to%20update%20what%20you%20know%20about%20my%20home%20environment.",
+                  kind: "text",
+                  label: "Messages",
+                },
+                {
+                  href: "https://t.me/withmurph_bot",
+                  kind: "telegram",
+                  label: "Telegram",
+                },
+              ]}
             />
           </Section>
         </div>
@@ -890,7 +897,7 @@ export function ComponentsContent() {
               built only from facts Murph still does not know.
             </p>
             <EnvironmentCaptureCard
-              contactAction={null}
+              contactOptions={[]}
               coverage={70}
               known={21}
               script={DESIGN_ENVIRONMENT_GAP_SCRIPT}
